Rewrite: wait during the uploading time, not more
Previous files have to be deleted (GDrive keeps all previous versions)

Then, we should also check files at the end of the list (even if it the
file should not be at the end of this list).

When selecting a file, there is first a dialogue, then we should check
the text next to our new file.

Upload a second file only if we have time to do that.